moved around some things and made the upsell look like the data studio ones
