+++ enterprise/frontend/src/metabase-enterprise/transforms/upsells/pages/TransformsUpsellPage/TransformsUpsellPage.tsx
@@ -3,10 +3,10 @@ import { useState } from "react";
 import { t } from "ttag";
 
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
+import { DottedBackground } from "metabase/common/components/upsells/components/DottedBackground";
+import { LineDecorator } from "metabase/common/components/upsells/components/LineDecorator";
 import { DataStudioBreadcrumbs } from "metabase/data-studio/common/components/DataStudioBreadcrumbs";
 import { PaneHeader } from "metabase/data-studio/common/components/PaneHeader";
-import { DottedBackground } from "metabase/data-studio/upsells/components/DottedBackground";
-import { LineDecorator } from "metabase/data-studio/upsells/components/LineDecorator";
 import type { BillingPeriod } from "metabase/data-studio/upsells/types";
 import { getStoreUsers } from "metabase/selectors/store-users";
 import {
