+++ frontend/src/metabase/common/components/upsells/components/DottedBackground/DottedBackground.tsx
