+++ frontend/src/metabase/common/components/upsells/components/DottedBackground/index.ts
