+++ frontend/src/metabase/common/components/upsells/components/LineDecorator/LineDecorator.module.css
