+++ frontend/src/metabase/common/components/upsells/components/LineDecorator/index.ts
