+++ frontend/src/metabase/common/components/upsells/components/UpsellCardContent.module.css
