+++ frontend/src/metabase/data-studio/upsells/pages/BaseUpsellPage.tsx
@@ -1,12 +1,11 @@
+import { DottedBackground } from "metabase/common/components/upsells/components/DottedBackground";
+import { LineDecorator } from "metabase/common/components/upsells/components/LineDecorator";
 import { useUpgradeAction } from "metabase/common/components/upsells/components/UpgradeModal";
+import { UpsellCardContent } from "metabase/common/components/upsells/components/UpsellCardContent";
 import { DATA_STUDIO_UPGRADE_URL } from "metabase/common/components/upsells/constants";
 import { DataStudioBreadcrumbs } from "metabase/data-studio/common/components/DataStudioBreadcrumbs";
 import { PaneHeader } from "metabase/data-studio/common/components/PaneHeader";
-import { UpsellCardContent } from "metabase/data-studio/upsells/components/UpsellCardContent";
 import { Stack } from "metabase/ui";
-
-import { DottedBackground } from "../components/DottedBackground";
-import { LineDecorator } from "../components/LineDecorator";
 
 import S from "./BaseUpsellPage.module.css";
